Split test_syscheck_audit into multiple tests.
