+++ src/unit_tests/syscheckd/whodata/test_win_whodata.c
